requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
iso8601==2.1.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pypika-tortoise==0.6.2
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
